+++ main.tf
@@ -11,7 +11,7 @@ locals {
 
 # Kubernetes cluster:
 module "kubernetes_cluster_operations" {
-  source = "github.com/kentrikos/terraform-aws-eks?ref=2.1.0"
+  source = "github.com/kentrikos/terraform-aws-eks?ref=2.2.0"
 
   cluster_prefix                = "${local.cluster_name}"
   region                        = "${var.region}"
@@ -28,6 +28,14 @@ module "kubernetes_cluster_operations" {
   install_helm                  = "${var.k8s_install_helm}"
   ingress_deploy                = false
   allowed_worker_ssh_cidrs      = "${var.k8s_allowed_worker_ssh_cidrs}"
+
+  map_roles            = "${var.map_roles}"
+  map_roles_count      = "${var.map_roles_count}"
+  map_users            = "${var.map_users}"
+  map_users_count      = "${var.map_users_count}"
+  map_accounts         = "${var.map_accounts}"
+  map_accounts_count   = "${var.map_accounts_count}"
+  enable_default_roles = "${var.enable_default_roles}"
 
   tags = "${local.common_tags}"
 }