+++ variables.tf
@@ -99,3 +99,44 @@ variable "k8s_allowed_worker_ssh_cidrs" {
   description = "List of CIDR ranges to allow SSH access into worker nodes"
   default     = []
 }
+
+variable "map_accounts" {
+  description = "Additional AWS account numbers to add to the aws-auth configmap. See terraform-aws-modules-eks examples/basic/variables.tf for example format."
+  type        = "list"
+  default     = []
+}
+
+variable "map_accounts_count" {
+  description = "The count of accounts in the map_accounts list."
+  type        = "string"
+  default     = 0
+}
+
+variable "map_roles" {
+  description = "Additional IAM roles to add to the aws-auth configmap. See terraform-aws-modules-eks examples/basic/variables.tf for example format."
+  type        = "list"
+  default     = []
+}
+
+variable "map_roles_count" {
+  description = "The count of roles in the map_roles list."
+  type        = "string"
+  default     = 0
+}
+
+variable "map_users" {
+  description = "Additional IAM users to add to the aws-auth configmap. See terraform-aws-modules-eksexamples/basic/variables.tf for example format."
+  type        = "list"
+  default     = []
+}
+
+variable "map_users_count" {
+  description = "The count of roles in the map_users list."
+  type        = "string"
+  default     = 0
+}
+
+variable "enable_default_roles" {
+  description = "Enable creation of default roles to assume"
+  default     = true
+}